moved ScanParametersFactory class to a separate package. Renamed 'ScanPipelineCommands' to PluginParametersHandlerTest for better readability.
